+++ friends.js
@@ -143,3 +143,9 @@ document.addEventListener("DOMContentLoaded", () => {
   spawnButton.addEventListener("keyup", () => { spawnButtonHeld = false })
   spawnButton.addEventListener("touchend", () => { spawnButtonHeld = false })
 })
+
+// Change emoji when new emoji selected
+document.querySelector('emoji-picker')
+  .addEventListener('emoji-click', (event) => {
+    document.body.style.setProperty("--selectedFriend", `"${event.detail.unicode}"`)
+  })